Refactoring AnimatedPresenting and AnimatedTransitioning protocols
